refactor: extract stop-processing.ts from daemon/main.ts
Move the stop-event processing pipeline from main() into a new
createStopProcessor() factory module. Exposes handleStopRoute,
clearSession, getActiveProcessing, and triggerTitleSummary.
Also exports enrichTurnsWithToolMetadata as a standalone pure function.
